Fix tests and remove legacy event builder.
