+++ lib/core/event_builder/index.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016-2017, Optimizely
+ * Copyright 2017, Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,14 @@
  * limitations under the License.
  */
 var fns = require('../../utils/fns');
-var lodashForOwn = require('lodash/forOwn');
+var eventTagUtils = require('../../utils/event_tag_utils');
 var projectConfig = require('../project_config');
+var uuid = require('uuid');
 
-var CONVERSION_ENDPOINT = 'https://logx.optimizely.com/log/event';
-var IMPRESSION_ENDPOINT = 'https://logx.optimizely.com/log/decision';
+var ACTIVATE_EVENT_KEY = 'campaign_activated';
 var CUSTOM_ATTRIBUTE_FEATURE_TYPE = 'custom';
+var ENDPOINT = 'https://logx.optimizely.com/v1/events';
 var HTTP_VERB = 'POST';
-var REVENUE_EVENT_METRIC_NAME = 'revenue';
-
 
 /**
  * Get params which are used same in both conversion and impression events
@@ -36,34 +35,34 @@ var REVENUE_EVENT_METRIC_NAME = 'revenue';
 function getCommonEventParams(options) {
   var attributes = options.attributes;
   var configObj = options.configObj;
-  var userId = options.userId;
 
-  var commonParams = {
-    accountId: configObj.accountId,
-    projectId: configObj.projectId,
-    revision: configObj.revision,
-    visitorId: userId,
-    timestamp: Math.round(new Date().getTime()),
-    isGlobalHoldback: false,
-    userFeatures: [],
-    // Used for tracking SDK type and version
-    clientEngine: options.clientEngine,
-    clientVersion: options.clientVersion,
+  var visitor = {
+    snapshots: [],
+    visitor_id: options.userId,
+    attributes: []
   };
 
-  for (var attributeKey in attributes) {
-    var attributeId = projectConfig.getAttributeId(options.configObj, attributeKey);
+  var commonParams = {
+    account_id: configObj.accountId,
+    project_id: configObj.projectId,
+    visitors: [visitor],
+    revision: configObj.revision,
+    client_name: options.clientEngine,
+    client_version: options.clientVersion,
+  };
+
+  fns.forOwn(attributes, function(attributeKey, attributeValue){
+    var attributeId = projectConfig.getAttributeId(options.configObj, attributeValue);
     if (attributeId) {
       var feature = {
-        id: attributeId,
-        name: attributeKey,
+        entity_id: attributeId,
+        key: attributeValue,
         type: CUSTOM_ATTRIBUTE_FEATURE_TYPE,
-        value: attributes[attributeKey],
-        shouldIndex: true,
+        value: attributes[attributeValue],
       };
-      commonParams.userFeatures.push(feature);
+      commonParams.visitors[0].attributes.push(feature);
     }
-  }
+  });
   return commonParams;
 }
 
@@ -76,13 +75,19 @@ function getCommonEventParams(options) {
  */
 function getImpressionEventParams(configObj, experimentId, variationId) {
   var impressionEventParams = {
-    layerId: projectConfig.getLayerId(configObj, experimentId),
-    decision: {
-      isLayerHoldback: false,
-      experimentId: experimentId,
-      variationId: variationId,
-    },
-  };
+      decisions: [{
+        campaign_id: projectConfig.getLayerId(configObj, experimentId),
+        experiment_id: experimentId,
+        variation_id: variationId,
+      }],
+      events: [{
+        entity_id: projectConfig.getLayerId(configObj, experimentId),
+        timestamp: Math.round(new Date().getTime()),
+        key: ACTIVATE_EVENT_KEY,
+        uuid: uuid.v4()
+      }]
+
+    };
   return impressionEventParams;
 }
 
@@ -92,57 +97,53 @@ function getImpressionEventParams(configObj, experimentId, variationId) {
  * @param  {string} eventKey                  Event key representing the event which needs to be recorded
  * @param  {Object} eventTags                 Values associated with the event.
  * @param  {Array}  experimentsToVariationMap Map of experiment IDs to bucketed variation IDs
+ * @param  {Object} logger                    Logger object
  * @return {Object}                           Conversion event params
  */
-function getConversionEventParams(configObj, eventKey, eventTags, experimentsToVariationMap) {
-  var conversionEventParams = {
-    eventEntityId: projectConfig.getEventId(configObj, eventKey),
-    eventName: eventKey,
-    eventFeatures: [],
-    layerStates: [],
-  };
+function getConversionEventParams(configObj, eventKey, eventTags, experimentsToVariationMap, logger) {
 
-  lodashForOwn(experimentsToVariationMap, function(variationId, experimentId) {
-    var layerState = {
-      layerId: projectConfig.getLayerId(configObj, experimentId),
-      revision: configObj.revision,
-      decision: {
-        isLayerHoldback: false,
-        experimentId: experimentId,
-        variationId: variationId,
-      },
-      actionTriggered: true,
+  var conversionEventParams = [];
+
+  fns.forOwn(experimentsToVariationMap, function(variationId, experimentId) {
+
+    var decision = {
+      decisions: [{
+        campaign_id: projectConfig.getLayerId(configObj, experimentId),
+        experiment_id: experimentId,
+        variation_id: variationId,
+      }],
+      events: [],
     };
 
-    conversionEventParams.layerStates.push(layerState);
-  });
+    var eventDict = {
+      entity_id: projectConfig.getEventId(configObj, eventKey),
+      timestamp: Math.round(new Date().getTime()),
+      uuid: uuid.v4(),
+      key: eventKey,
+    };
 
-  conversionEventParams.eventMetrics = [];
-
-  if (eventTags) {
-    fns.forEach(eventTags, function(eventTagValue, eventTagKey) {
-      if (eventTagKey === REVENUE_EVENT_METRIC_NAME) {
-        var revenueMetric = {
-          name: REVENUE_EVENT_METRIC_NAME,
-          value: eventTagValue,
-        };
-        conversionEventParams.eventMetrics.push(revenueMetric);
+    if (eventTags) {
+      var revenue = eventTagUtils.getRevenueValue(eventTags, logger);
+      if (revenue) {
+        eventDict['revenue'] = revenue;
       }
-      var eventFeature = {
-        name: eventTagKey,
-        type: CUSTOM_ATTRIBUTE_FEATURE_TYPE,
-        value: eventTagValue,
-        shouldIndex: false,
-      };
-      conversionEventParams.eventFeatures.push(eventFeature);
-    });
-  }
+
+      var eventValue = eventTagUtils.getEventValue(eventTags, logger);
+      if (eventValue) {
+        eventDict['value'] = eventValue;
+      }
+
+      eventDict['tags'] = eventTags;
+    }
+    decision.events = [eventDict];
+
+    conversionEventParams.push(decision);
+  });
 
   return conversionEventParams;
 }
 
 module.exports = {
-
   /**
    * Create impression event params to be sent to the logging endpoint
    * @param  {Object} options               Object containing values needed to build impression event
@@ -161,10 +162,13 @@ module.exports = {
     };
 
     var commonParams = getCommonEventParams(options);
-    impressionEvent.url = IMPRESSION_ENDPOINT;
+    impressionEvent.url = ENDPOINT;
 
     var impressionEventParams = getImpressionEventParams(options.configObj, options.experimentId, options.variationId);
-    impressionEvent.params = fns.assignIn(commonParams, impressionEventParams);
+    // combine Event params into visitor obj
+    commonParams.visitors[0].snapshots.push(impressionEventParams);
+
+    impressionEvent.params = commonParams;
 
     return impressionEvent;
   },
@@ -178,8 +182,9 @@ module.exports = {
    * @param  {Object} options.configObj                 Object representing project configuration, including datafile information and mappings for quick lookup
    * @param  {string} options.eventKey                  Event key representing the event which needs to be recorded
    * @param  {Object} options.eventTags                 Object with event-specific tags
-   * @param  {string} options.userId                    ID for user
    * @param  {Object} options.experimentsToVariationMap Map of experiment IDs to bucketed variation IDs
+   * @param  {Object} options.logger                    Logger object
+   * @param  {string} options.userId                    ID for user
    * @return {Object}                                   Params to be used in conversion event logging endpoint call
    */
   getConversionEvent: function(options) {
@@ -188,13 +193,17 @@ module.exports = {
     };
 
     var commonParams = getCommonEventParams(options);
-    conversionEvent.url = CONVERSION_ENDPOINT;
+    conversionEvent.url = ENDPOINT;
 
     var conversionEventParams = getConversionEventParams(options.configObj,
                                                          options.eventKey,
                                                          options.eventTags,
-                                                         options.experimentsToVariationMap);
-    conversionEvent.params = fns.assignIn(commonParams, conversionEventParams);
+                                                         options.experimentsToVariationMap,
+                                                         options.logger);
+
+    commonParams.visitors[0].snapshots = conversionEventParams;
+    conversionEvent.params = commonParams;
+
     return conversionEvent;
   },
 };
