+++ lib/core/event_builder/tests.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016-2017, Optimizely
+ * Copyright 2017, Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-var eventBuilder = require('./');
+var eventBuilder = require('./index_v3.js');
 var packageJSON = require('../../../package.json');
 var projectConfig = require('../project_config');
 var testData = require('../../../tests/test_data');
@@ -21,6 +21,8 @@ var testData = require('../../../tests/test_data');
 var chai = require('chai');
 var assert = chai.assert;
 var sinon = require('sinon');
+var uuid = require('uuid');
+
 
 describe('lib/core/event_builder', function() {
   describe('APIs', function() {
@@ -31,34 +33,44 @@ describe('lib/core/event_builder', function() {
     beforeEach(function() {
       configObj = projectConfig.createProjectConfig(testData.getTestProjectConfig());
       clock = sinon.useFakeTimers(new Date().getTime());
+      sinon.stub(uuid, 'v4').returns('a68cf1ad-0393-4e18-af87-efe8f01a7c9c');
     });
 
     afterEach(function() {
       clock.restore();
+      uuid.v4.restore();
     });
 
     describe('getImpressionEvent', function() {
       it('should create proper params for getImpressionEvent without attributes', function() {
+
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/decision',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            isGlobalHoldback: false,
-            projectId: '111001',
-            revision: '42',
-            decision: {
-              variationId: '111128',
-              isLayerHoldback: false,
-              experimentId: '111127',
-            },
-            layerId: '4',
-            accountId: '12001',
-            userFeatures: [],
-          },
+            'account_id': '12001',
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [],
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'timestamp': Math.round(new Date().getTime()),
+                  'entity_id': '4',
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'key': 'campaign_activated'
+                }]
+              }]
+            }],
+            'revision': '42',
+            'client_name': 'node-sdk',
+            'client_version': packageJSON.version
+          }
         };
 
         var eventOptions = {
@@ -77,35 +89,38 @@ describe('lib/core/event_builder', function() {
 
       it('should create proper params for getImpressionEvent with attributes as a string value', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/decision',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            isGlobalHoldback: false,
-            projectId: '111001',
-            revision: '42',
-            decision: {
-              variationId: '111128',
-              isLayerHoldback: false,
-              experimentId: '111127',
-            },
-            layerId: '4',
-            accountId: '12001',
-            userFeatures: [
-              {
-                id: '111094',
-                name: 'browser_type',
-                type: 'custom',
-                value: 'firefox',
-                shouldIndex: true,
-              },
-            ],
-          },
+            'account_id': '12001',
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [{
+                'entity_id': '111094',
+                'type': 'custom',
+                'value': 'firefox',
+                'key': 'browser_type'
+              }],
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'timestamp': Math.round(new Date().getTime()),
+                  'entity_id': '4',
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'key': 'campaign_activated'
+                }]
+              }]
+            }],
+            'revision': '42',
+            'client_name': 'node-sdk',
+            'client_version': packageJSON.version
+          }
         };
-
         var eventOptions = {
           attributes: {browser_type: 'firefox'},
           clientEngine: 'node-sdk',
@@ -123,33 +138,37 @@ describe('lib/core/event_builder', function() {
 
       it('should create proper params for getImpressionEvent with attributes as a false value', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/decision',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            isGlobalHoldback: false,
-            projectId: '111001',
-            revision: '42',
-            decision: {
-              variationId: '111128',
-              isLayerHoldback: false,
-              experimentId: '111127',
-            },
-            layerId: '4',
-            accountId: '12001',
-            userFeatures: [
-              {
-                id: '111094',
-                name: 'browser_type',
-                type: 'custom',
-                value: false,
-                shouldIndex: true,
-              },
-            ],
-          },
+            'account_id': '12001',
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [{
+                'entity_id': '111094',
+                'type': 'custom',
+                'value': false,
+                'key': 'browser_type'
+              }],
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'timestamp': Math.round(new Date().getTime()),
+                  'entity_id': '4',
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'key': 'campaign_activated'
+                }]
+              }]
+            }],
+            'revision': '42',
+            'client_name': 'node-sdk',
+            'client_version': packageJSON.version
+          }
         };
 
         var eventOptions = {
@@ -169,33 +188,37 @@ describe('lib/core/event_builder', function() {
 
       it('should create proper params for getImpressionEvent with attributes as a zero value', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/decision',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            isGlobalHoldback: false,
-            projectId: '111001',
-            revision: '42',
-            decision: {
-              variationId: '111128',
-              isLayerHoldback: false,
-              experimentId: '111127',
-            },
-            layerId: '4',
-            accountId: '12001',
-            userFeatures: [
-              {
-                id: '111094',
-                name: 'browser_type',
-                type: 'custom',
-                value: 0,
-                shouldIndex: true,
-              },
-            ],
-          },
+            'account_id': '12001',
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [{
+                'entity_id': '111094',
+                'type': 'custom',
+                'value': 0,
+                'key': 'browser_type'
+              }],
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'timestamp': Math.round(new Date().getTime()),
+                  'entity_id': '4',
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'key': 'campaign_activated'
+                }]
+              }]
+            }],
+            'revision': '42',
+            'client_name': 'node-sdk',
+            'client_version': packageJSON.version
+          }
         };
 
         var eventOptions = {
@@ -215,25 +238,32 @@ describe('lib/core/event_builder', function() {
 
       it('should not fill in userFeatures for getImpressionEvent when attribute is not in the datafile', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/decision',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            isGlobalHoldback: false,
-            projectId: '111001',
-            revision: '42',
-            decision: {
-              variationId: '111128',
-              isLayerHoldback: false,
-              experimentId: '111127',
-            },
-            layerId: '4',
-            accountId: '12001',
-            userFeatures: [],
-          },
+            'account_id': '12001',
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [],
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'timestamp': Math.round(new Date().getTime()),
+                  'entity_id': '4',
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'key': 'campaign_activated'
+                }]
+              }]
+            }],
+            'revision': '42',
+            'client_name': 'node-sdk',
+            'client_version': packageJSON.version
+          }
         };
 
         var eventOptions = {
@@ -253,35 +283,42 @@ describe('lib/core/event_builder', function() {
     });
 
     describe('getConversionEvent', function() {
+      var mockLogger;
+
+      beforeEach(function() {
+        mockLogger = {
+          log: sinon.stub(),
+        };
+      });
+
       it('should create proper params for getConversionEvent without attributes or event value', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
+            'account_id': '12001',
+            'project_id': '111001',
+            'visitors': [{
+              'visitor_id': 'testUser',
+              'attributes': [],
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'timestamp': Math.round(new Date().getTime()),
+                  'entity_id': '111095',
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'key': 'testEvent'
+                }]
+              }]
             }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [],
-            eventFeatures: [],
-            isGlobalHoldback: false,
-          },
+            'revision': '42',
+            'client_name': 'node-sdk',
+            'client_version': packageJSON.version
+          }
         };
 
         var eventOptions = {
@@ -289,6 +326,7 @@ describe('lib/core/event_builder', function() {
           clientVersion: packageJSON.version,
           configObj: configObj,
           eventKey: 'testEvent',
+          logger: mockLogger,
           userId: 'testUser',
           experimentsToVariationMap: { '111127': '111128' },
         };
@@ -300,39 +338,37 @@ describe('lib/core/event_builder', function() {
 
       it('should create proper params for getConversionEvent with attributes', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [{
-                id: '111094',
-                name: 'browser_type',
-                type: 'custom',
-                value: 'firefox',
-                shouldIndex: true,
+            'account_id': '12001',
+            'project_id': '111001',
+            'visitors': [{
+              'visitor_id': 'testUser',
+              'attributes': [{
+                'entity_id': '111094',
+                'type': 'custom',
+                'value': 'firefox',
+                'key': 'browser_type'
               }],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'timestamp': Math.round(new Date().getTime()),
+                  'entity_id': '111095',
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'key': 'testEvent'
+                }]
+              }]
             }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [],
-            eventFeatures: [],
-            isGlobalHoldback: false,
-          },
+            'revision': '42',
+            'client_name': 'node-sdk',
+            'client_version': packageJSON.version
+          }
         };
 
         var eventOptions = {
@@ -341,8 +377,9 @@ describe('lib/core/event_builder', function() {
           clientVersion: packageJSON.version,
           configObj: configObj,
           eventKey: 'testEvent',
-          userId: 'testUser',
           experimentsToVariationMap: { '111127': '111128' },
+          logger: mockLogger,
+          userId: 'testUser',
         };
 
         var actualParams = eventBuilder.getConversionEvent(eventOptions);
@@ -352,41 +389,36 @@ describe('lib/core/event_builder', function() {
 
       it('should create proper params for getConversionEvent with event value', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
+            'client_version': packageJSON.version,
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [],
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'tags': {
+                    'revenue': 4200
+                  },
+                  'timestamp': Math.round(new Date().getTime()),
+                  'revenue': 4200,
+                  'key': 'testEvent',
+                  'entity_id': '111095'
+                }]
+              }]
             }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [{
-              name: 'revenue',
-              value: 4200,
-            }],
-            eventFeatures: [{
-              "name": "revenue",
-              "type": 'custom',
-              "value": 4200,
-              "shouldIndex": false,
-            }],
-            isGlobalHoldback: false,
-          },
+            'account_id': '12001',
+            'client_name': 'node-sdk',
+            'revision': '42'
+          }
         };
 
         var eventOptions = {
@@ -397,8 +429,9 @@ describe('lib/core/event_builder', function() {
           eventTags: {
             revenue: 4200,
           },
-          userId: 'testUser',
           experimentsToVariationMap: { '111127': '111128' },
+          logger: mockLogger,
+          userId: 'testUser',
         };
 
         var actualParams = eventBuilder.getConversionEvent(eventOptions);
@@ -408,47 +441,41 @@ describe('lib/core/event_builder', function() {
 
       it('should create proper params for getConversionEvent with attributes and event value', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [{
-                id: '111094',
-                name: 'browser_type',
-                type: 'custom',
-                value: 'firefox',
-                shouldIndex: true,
+            'client_version': packageJSON.version,
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [{
+                'entity_id': '111094',
+                'type': 'custom',
+                'value': 'firefox',
+                'key': 'browser_type'
               }],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'tags': {
+                    'revenue': 4200
+                  },
+                  'timestamp': Math.round(new Date().getTime()),
+                  'revenue': 4200,
+                  'key': 'testEvent',
+                  'entity_id': '111095'
+                }]
+              }]
             }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [{
-              name: 'revenue',
-              value: 4200,
-            }],
-            eventFeatures: [{
-              "name": "revenue",
-              "type": 'custom',
-              "value": 4200,
-              "shouldIndex": false,
-            }],
-            isGlobalHoldback: false,
-          },
+            'account_id': '12001',
+            'client_name': 'node-sdk',
+            'revision': '42'
+          }
         };
 
         var eventOptions = {
@@ -460,8 +487,9 @@ describe('lib/core/event_builder', function() {
           eventTags: {
             revenue: 4200
           },
-          userId: 'testUser',
           experimentsToVariationMap: { '111127': '111128' },
+          logger: mockLogger,
+          userId: 'testUser',
         };
 
         var actualParams = eventBuilder.getConversionEvent(eventOptions);
@@ -471,47 +499,41 @@ describe('lib/core/event_builder', function() {
 
       it('should create proper params for getConversionEvent with attributes and revenue', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [{
-                id: '111094',
-                name: 'browser_type',
-                type: 'custom',
-                value: 'firefox',
-                shouldIndex: true,
+            'client_version': packageJSON.version,
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [{
+                'entity_id': '111094',
+                'type': 'custom',
+                'value': 'firefox',
+                'key': 'browser_type'
               }],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'tags': {
+                    'revenue': 4200
+                  },
+                  'timestamp': Math.round(new Date().getTime()),
+                  'revenue': 4200,
+                  'key': 'testEvent',
+                  'entity_id': '111095'
+                }]
+              }]
             }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [{
-              name: 'revenue',
-              value: 4200,
-            }],
-            eventFeatures: [{
-              "name": "revenue",
-              "type": 'custom',
-              "value": 4200,
-              "shouldIndex": false,
-            }],
-            isGlobalHoldback: false,
-          },
+            'account_id': '12001',
+            'client_name': 'node-sdk',
+            'revision': '42'
+          }
         };
 
         var eventOptions = {
@@ -523,123 +545,9 @@ describe('lib/core/event_builder', function() {
           eventTags: {
             revenue: 4200,
           },
-          userId: 'testUser',
           experimentsToVariationMap: { '111127': '111128' },
-        };
-
-        var actualParams = eventBuilder.getConversionEvent(eventOptions);
-
-        assert.deepEqual(actualParams, expectedParams);
-      });
-
-      it('should create proper params for getConversionEvent with event tags', function() {
-        var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
-          httpVerb: 'POST',
-          params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
-            }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [],
-            eventFeatures: [{
-              "name": "non-revenue",
-              "type": 'custom',
-              "value": "cool",
-              "shouldIndex": false,
-            }],
-            isGlobalHoldback: false,
-          },
-        };
-
-        var eventOptions = {
-          clientEngine: 'node-sdk',
-          clientVersion: packageJSON.version,
-          configObj: configObj,
-          eventKey: 'testEvent',
-          eventTags: {
-            'non-revenue': 'cool',
-          },
+          logger: mockLogger,
           userId: 'testUser',
-          experimentsToVariationMap: { '111127': '111128' },
-        };
-
-        var actualParams = eventBuilder.getConversionEvent(eventOptions);
-
-        assert.deepEqual(actualParams, expectedParams);
-      });
-
-      it('should create proper params for getConversionEvent with event tags and revenue', function() {
-        var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
-          httpVerb: 'POST',
-          params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
-            }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [{
-              name: 'revenue',
-              value: 4200,
-            }],
-            eventFeatures: [{
-              "name": "revenue",
-              "type": 'custom',
-              "value": 4200,
-              "shouldIndex": false,
-            }, {
-              "name": "non-revenue",
-              "type": 'custom',
-              "value": "cool",
-              "shouldIndex": false,
-            }],
-            isGlobalHoldback: false,
-          },
-        };
-
-        var eventOptions = {
-          clientEngine: 'node-sdk',
-          clientVersion: packageJSON.version,
-          configObj: configObj,
-          eventKey: 'testEvent',
-          eventTags: {
-            'revenue': 4200,
-            'non-revenue': 'cool',
-          },
-          userId: 'testUser',
-          experimentsToVariationMap: { '111127': '111128' },
         };
 
         var actualParams = eventBuilder.getConversionEvent(eventOptions);
@@ -649,33 +557,32 @@ describe('lib/core/event_builder', function() {
 
       it('should not fill in userFeatures for getConversion when attribute is not in the datafile', function() {
         var expectedParams = {
-          url: 'https://logx.optimizely.com/log/event',
+          url: 'https://logx.optimizely.com/v1/events',
           httpVerb: 'POST',
           params: {
-            visitorId: 'testUser',
-            timestamp: Math.round(new Date().getTime()),
-            clientEngine: 'node-sdk',
-            clientVersion: packageJSON.version,
-            projectId: '111001',
-            revision: '42',
-            accountId: '12001',
-            userFeatures: [],
-            layerStates: [{
-              layerId: '4',
-              revision: '42',
-              decision: {
-                variationId: '111128',
-                isLayerHoldback: false,
-                experimentId: '111127',
-              },
-              actionTriggered: true,
+            'client_version': packageJSON.version,
+            'project_id': '111001',
+            'visitors': [{
+              'attributes': [],
+              'visitor_id': 'testUser',
+              'snapshots': [{
+                'decisions': [{
+                  'variation_id': '111128',
+                  'experiment_id': '111127',
+                  'campaign_id': '4'
+                }],
+                'events': [{
+                  'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                  'timestamp': Math.round(new Date().getTime()),
+                  'key': 'testEvent',
+                  'entity_id': '111095'
+                }]
+              }]
             }],
-            eventEntityId: '111095',
-            eventName: 'testEvent',
-            eventMetrics: [],
-            eventFeatures: [],
-            isGlobalHoldback: false,
-          },
+            'account_id': '12001',
+            'client_name': 'node-sdk',
+            'revision': '42'
+          }
         };
 
         var eventOptions = {
@@ -684,13 +591,289 @@ describe('lib/core/event_builder', function() {
           clientVersion: packageJSON.version,
           configObj: configObj,
           eventKey: 'testEvent',
-          userId: 'testUser',
           experimentsToVariationMap: { '111127': '111128' },
+          logger: mockLogger,
+          userId: 'testUser',
         };
 
         var actualParams = eventBuilder.getConversionEvent(eventOptions);
 
         assert.deepEqual(actualParams, expectedParams);
+      });
+
+      describe('and event tags are passed it', function() {
+        it('should create proper params for getConversionEvent with event tags', function() {
+          var expectedParams = {
+            url: 'https://logx.optimizely.com/v1/events',
+            httpVerb: 'POST',
+            params: {
+              'client_version': packageJSON.version,
+              'project_id': '111001',
+              'visitors': [{
+                'attributes': [],
+                'visitor_id': 'testUser',
+                'snapshots': [{
+                  'decisions': [{
+                    'variation_id': '111128',
+                    'experiment_id': '111127',
+                    'campaign_id': '4'
+                  }],
+                  'events': [{
+                    'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                    'tags': {
+                      'non-revenue': 'cool',
+                    },
+                    'timestamp': Math.round(new Date().getTime()),
+                    'key': 'testEvent',
+                    'entity_id': '111095'
+                  }]
+                }]
+              }],
+              'account_id': '12001',
+              'client_name': 'node-sdk',
+              'revision': '42'
+            }
+          };
+
+          var eventOptions = {
+            clientEngine: 'node-sdk',
+            clientVersion: packageJSON.version,
+            configObj: configObj,
+            eventKey: 'testEvent',
+            eventTags: {
+              'non-revenue': 'cool',
+            },
+            experimentsToVariationMap: { '111127': '111128' },
+            logger: mockLogger,
+            userId: 'testUser',
+          };
+
+          var actualParams = eventBuilder.getConversionEvent(eventOptions);
+
+          assert.deepEqual(actualParams, expectedParams);
+        });
+
+        describe('and the event tags contain an entry for "revenue"', function() {
+          it('should include the revenue value in the event object', function() {
+            var expectedParams = {
+              url: 'https://logx.optimizely.com/v1/events',
+              httpVerb: 'POST',
+              params: {
+                'client_version': packageJSON.version,
+                'project_id': '111001',
+                'visitors': [{
+                  'attributes': [],
+                  'visitor_id': 'testUser',
+                  'snapshots': [{
+                    'decisions': [{
+                      'variation_id': '111128',
+                      'experiment_id': '111127',
+                      'campaign_id': '4'
+                    }],
+                    'events': [{
+                      'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                      'tags': {
+                        'non-revenue': 'cool',
+                        'revenue': 4200
+                      },
+                      'timestamp': Math.round(new Date().getTime()),
+                      'revenue': 4200,
+                      'key': 'testEvent',
+                      'entity_id': '111095'
+                    }]
+                  }]
+                }],
+                'account_id': '12001',
+                'client_name': 'node-sdk',
+                'revision': '42'
+              }
+            };
+
+            var eventOptions = {
+              clientEngine: 'node-sdk',
+              clientVersion: packageJSON.version,
+              configObj: configObj,
+              eventKey: 'testEvent',
+              eventTags: {
+                'revenue': 4200,
+                'non-revenue': 'cool',
+              },
+              experimentsToVariationMap: { '111127': '111128' },
+              logger: mockLogger,
+              userId: 'testUser',
+            };
+
+            var actualParams = eventBuilder.getConversionEvent(eventOptions);
+
+            assert.deepEqual(actualParams, expectedParams);
+          });
+
+          describe('and the revenue value is invalid', function() {
+            it('should not include the revenue value in the event object', function() {
+              var expectedParams = {
+                url: 'https://logx.optimizely.com/v1/events',
+                httpVerb: 'POST',
+                params: {
+                  'client_version': packageJSON.version,
+                  'project_id': '111001',
+                  'visitors': [{
+                    'attributes': [],
+                    'visitor_id': 'testUser',
+                    'snapshots': [{
+                      'decisions': [{
+                        'variation_id': '111128',
+                        'experiment_id': '111127',
+                        'campaign_id': '4'
+                      }],
+                      'events': [{
+                        'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                        'tags': {
+                          'non-revenue': 'cool',
+                          'revenue': 'invalid revenue'
+                        },
+                        'timestamp': Math.round(new Date().getTime()),
+                        'key': 'testEvent',
+                        'entity_id': '111095'
+                      }]
+                    }]
+                  }],
+                  'account_id': '12001',
+                  'client_name': 'node-sdk',
+                  'revision': '42'
+                }
+              };
+
+              var eventOptions = {
+                clientEngine: 'node-sdk',
+                clientVersion: packageJSON.version,
+                configObj: configObj,
+                eventKey: 'testEvent',
+                eventTags: {
+                  'revenue': 'invalid revenue',
+                  'non-revenue': 'cool',
+                },
+                experimentsToVariationMap: { '111127': '111128' },
+                logger: mockLogger,
+                userId: 'testUser',
+              };
+
+              var actualParams = eventBuilder.getConversionEvent(eventOptions);
+
+              assert.deepEqual(actualParams, expectedParams);
+            });
+          });
+        });
+
+        describe('and the event tags contain an entry for "value"', function() {
+          it('should include the event value in the event object', function() {
+            var expectedParams = {
+              url: 'https://logx.optimizely.com/v1/events',
+              httpVerb: 'POST',
+              params: {
+                'client_version': packageJSON.version,
+                'project_id': '111001',
+                'visitors': [{
+                  'attributes': [],
+                  'visitor_id': 'testUser',
+                  'snapshots': [{
+                    'decisions': [{
+                      'variation_id': '111128',
+                      'experiment_id': '111127',
+                      'campaign_id': '4'
+                    }],
+                    'events': [{
+                      'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                      'tags': {
+                        'non-revenue': 'cool',
+                        'value': '13.37'
+                      },
+                      'timestamp': Math.round(new Date().getTime()),
+                      'value': 13.37,
+                      'key': 'testEvent',
+                      'entity_id': '111095'
+                    }]
+                  }]
+                }],
+                'account_id': '12001',
+                'client_name': 'node-sdk',
+                'revision': '42'
+              }
+            };
+
+            var eventOptions = {
+              clientEngine: 'node-sdk',
+              clientVersion: packageJSON.version,
+              configObj: configObj,
+              eventKey: 'testEvent',
+              eventTags: {
+                'value': '13.37',
+                'non-revenue': 'cool',
+              },
+              experimentsToVariationMap: { '111127': '111128' },
+              logger: mockLogger,
+              userId: 'testUser',
+            };
+
+            var actualParams = eventBuilder.getConversionEvent(eventOptions);
+
+            assert.deepEqual(actualParams, expectedParams);
+          });
+
+          describe('and the event value is invalid', function() {
+            it('should not include the event value in the event object', function() {
+              var expectedParams = {
+                url: 'https://logx.optimizely.com/v1/events',
+                httpVerb: 'POST',
+                params: {
+                  'client_version': packageJSON.version,
+                  'project_id': '111001',
+                  'visitors': [{
+                    'attributes': [],
+                    'visitor_id': 'testUser',
+                    'snapshots': [{
+                      'decisions': [{
+                        'variation_id': '111128',
+                        'experiment_id': '111127',
+                        'campaign_id': '4'
+                      }],
+                      'events': [{
+                        'uuid': 'a68cf1ad-0393-4e18-af87-efe8f01a7c9c',
+                        'tags': {
+                          'non-revenue': 'cool',
+                          'value': 'invalid value'
+                        },
+                        'timestamp': Math.round(new Date().getTime()),
+                        'key': 'testEvent',
+                        'entity_id': '111095'
+                      }]
+                    }]
+                  }],
+                  'account_id': '12001',
+                  'client_name': 'node-sdk',
+                  'revision': '42'
+                }
+              };
+
+              var eventOptions = {
+                clientEngine: 'node-sdk',
+                clientVersion: packageJSON.version,
+                configObj: configObj,
+                eventKey: 'testEvent',
+                eventTags: {
+                  'value': 'invalid value',
+                  'non-revenue': 'cool',
+                },
+                experimentsToVariationMap: { '111127': '111128' },
+                logger: mockLogger,
+                userId: 'testUser',
+              };
+
+              var actualParams = eventBuilder.getConversionEvent(eventOptions);
+
+              assert.deepEqual(actualParams, expectedParams);
+            });
+          });
+        });
       });
     });
   });
