+++ lib/optimizely/index.js
@@ -18,7 +18,7 @@ var fns = require('../utils/fns');
 var attributesValidator = require('../utils/attributes_validator');
 var decisionService = require('../core/decision_service');
 var enums = require('../utils/enums');
-var eventBuilder = require('../core/event_builder/index_v3.js');
+var eventBuilder = require('../core/event_builder/index.js');
 var eventTagsValidator = require('../utils/event_tags_validator');
 var projectConfig = require('../core/project_config');
 var projectConfigSchema = require('./project_config_schema');
